Refactor : FeignClient
