+++ src/main/java/com/spacestar/back/feignClient/dto/res/ProfileResDto.java
@@ -1,4 +1,4 @@
-package com.spacestar.back.quickmatching.dto.res;
+package com.spacestar.back.feignClient.dto.res;
 
 import lombok.Getter;
