+++ src/main/java/com/spacestar/back/quickmatching/service/QuickMatchingServiceImpl.java
@@ -6,12 +6,12 @@ import com.spacestar.back.global.ResponseEntity;
 import com.spacestar.back.global.ResponseStatus;
 import com.spacestar.back.quickmatching.converter.QuickMatchingConverter;
 import com.spacestar.back.quickmatching.domain.QuickMatching;
-import com.spacestar.back.quickmatching.dto.QuickMatchingEnterReqDto;
-import com.spacestar.back.quickmatching.dto.QuickMatchingResDto;
-import com.spacestar.back.quickmatching.dto.res.AuthResDto;
-import com.spacestar.back.quickmatching.dto.res.ProfileResDto;
-import com.spacestar.back.quickmatching.feignClient.AuthClient;
-import com.spacestar.back.quickmatching.feignClient.ProfileClient;
+import com.spacestar.back.quickmatching.dto.req.QuickMatchingEnterReqDto;
+import com.spacestar.back.quickmatching.dto.res.QuickMatchingResDto;
+import com.spacestar.back.feignClient.dto.res.AuthResDto;
+import com.spacestar.back.feignClient.dto.res.ProfileResDto;
+import com.spacestar.back.feignClient.AuthClient;
+import com.spacestar.back.feignClient.ProfileClient;
 import com.spacestar.back.quickmatching.repository.MatchingScoresRepository;
 import com.spacestar.back.quickmatching.repository.QuickMatchingRepository;
 import lombok.RequiredArgsConstructor;
