+++ src/main/java/com/spacestar/back/quickmatching/vo/req/QuickMatchingEnterReqVo.java
@@ -1,4 +1,4 @@
-package com.spacestar.back.quickmatching.vo;
+package com.spacestar.back.quickmatching.vo.req;
 
 import lombok.Getter;
