+++ src/main/java/com/spacestar/back/quickmatching/vo/res/QuickMatchingResVo.java
@@ -1,4 +1,4 @@
-package com.spacestar.back.quickmatching.vo;
+package com.spacestar.back.quickmatching.vo.res;
 
 import lombok.Getter;
